backend/requirements.txt
anyascii==0.3.3
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
defusedxml==0.7.1
easyocr==1.7.2
filelock==3.19.1
fsspec==2025.7.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
langdetect==1.0.9
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnx==1.18.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
protobuf==6.32.0
pyclipper==1.3.0.post6
PyMuPDF==1.26.3
pypdfium2==4.30.0
pytesseract==0.3.13
python-bidi==0.6.6
python-doctr==1.0.0
python-dotenv==1.1.1
python-telegram-bot==22.3
PyYAML==6.0.2
pyzbar==0.1.9
RapidFuzz==3.13.0
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tifffile==2025.6.11
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.14.1
urllib3==2.5.0
validators==0.35.0
